preprocessing-pipeline/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Text processing
nltk>=3.8.0
textstat>=0.7.3
langdetect>=1.0.9

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0

# Optional: For future Gutenberg API integration
requests>=2.31.0